refactor : 카드를 뽑고 리스트에서 제거하는 행위를 Deck.kt으로 위임
