+++ src/main/kotlin/blackjack/deck/Cards.kt
@@ -5,9 +5,10 @@ import blackjack.card.Card
 data class Cards(
     var cards: List<Card>,
 ) {
-    fun draw(): Pair<Card, List<Card>> {
-        require(cards.isNotEmpty()) { "카드 목록이 비어있습니다." }
-        return (cards.first() to cards.drop(1))
+    fun draw() = cards.first()
+
+    fun discard(card: Card) {
+        cards = cards.filterNot { it == card }
     }
 
     fun isNotEmpty() = cards.isNotEmpty()
